pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bharat-resto-mvp"
version = "0.1.0"
description = "End-to-end pipeline for discovering and extracting Indian restaurant information"
authors = [
    {name = "Bharat Resto Team", email = "team@bharatresto.local"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["restaurants", "india", "data-extraction", "osm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core dependencies
    "httpx>=0.24.0",
    "requests-cache>=1.1.0",
    "lxml>=4.9.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.0",
    "pdfminer.six>=20221105",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "geopy>=2.3.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.1.0",
    "python-slugify>=8.0.0",
    "rapidfuzz>=3.0.0",
    "geohash2>=1.1",
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
]

[project.scripts]
bharat-resto = "app.cli:main"

[project.urls]
Homepage = "https://github.com/bharatresto/mvp"
Repository = "https://github.com/bharatresto/mvp"
Issues = "https://github.com/bharatresto/mvp/issues"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v"
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by black)
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
